Refactored code to change typescript files to javascript files, renamed some naming conventions for less ambiguity
