--- frontend/src/components/Register.tsx
@@ -1,102 +1,0 @@
-import { useNavigate } from "react-router-dom";
-import { Button, TextField, Box, Typography, Alert } from "@mui/material";
-import { useState } from "react";
-import AuthService from "../services/authService";
-
-function Register() {
-    const [formData, setFormData] = useState({ name: "", email: "", password: "" });
-    const [renderMessage, setRenderMessage] = useState(false);
-    const [message, setMessage] = useState("");
-    const [messageSeverity, setMessageSeverity] = useState<"error" | "warning" | "info" | "success">("warning");
-
-    const navigate = useNavigate();
-
-    const handleRegister = (e: { preventDefault: () => void; }) => {
-        e.preventDefault();
-        setRenderMessage(false);
-        setMessage("")
-
-        AuthService.register(formData.name, formData.email, formData.password)
-        .then(() => {setMessage("Account registration successful");
-        setRenderMessage(true);
-        setMessageSeverity("success")
-        })
-        .catch((error: { response: { data: { message: any; }; }; message: any; toString: () => any; }) => {
-            const errorMessage = (error.response && error.response.data && error.response.data.message) || error.message || error.toString() 
-            || "Login failed. Please try again.";
-            setMessage(errorMessage);
-            setRenderMessage(true);
-            setMessageSeverity("warning");
-        })
-    };
-
-    const handleFormChange = (e: { target: { name: string; value: string; }; }) => {
-        setFormData({ ...formData, [e.target.name]: e.target.value });
-        console.log(formData);
-    }
-
-    return (
-        <Box sx={{
-            height: "100vh",
-            display: "flex",
-            justifyContent: "center",
-            alignItems: "center",
-        }}>
-            <form onSubmit={handleRegister}>
-            <div className="Register">
-                <Box sx={{
-                    width: 300,
-                    padding: 4,
-                    display: "flex",
-                    flexDirection: "column",
-                    gap: 3,
-                    boxShadow: 3,
-                    borderRadius: 2,
-                    backgroundColor: "#fafafa",
-                }}>
-                    {renderMessage && (<Alert variant="filled" severity={messageSeverity}> {message} </Alert>)}
-                    <Typography variant="h3" align="center">
-                        Register
-                    </Typography>
-                    <TextField
-                        id="filled-search"
-                        label="Name"
-                        name="name"
-                        type="search"
-                        variant="filled"
-                        onChange={handleFormChange}
-                        required
-                    />
-                    <TextField
-                        id="filled-search"
-                        label="Email"
-                        name="email"
-                        type="search"
-                        variant="filled"
-                        onChange={handleFormChange}
-                        required
-                    />
-                    <TextField
-                        id="filled-password-input"
-                        name="password"
-                        label="Password"
-                        type="password"
-                        autoComplete="current-password"
-                        variant="filled"
-                        onChange={handleFormChange}
-                        required
-                    />
-                    <Button variant="contained" type = "submit">Submit</Button>
-                    <a href="/">
-                    <Typography variant="h6" align="center">
-                        Login
-                    </Typography>
-                    </a>
-                </Box>
-            </div>
-            </form>
-        </Box>
-    )
-}
-
-export default Register;
