+++ frontend/src/index.jsx
@@ -2,7 +2,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 
 const body = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+  document.getElementById('root')
 );
 body.render(
     <App />
